+++ src/parser/link_url/parse_link.rs
@@ -82,8 +82,11 @@ fn is_other_scheme(c: char) -> bool {
     matches!(c, '+' | '-' | '.')
 }
 
+/**
+ * allowed chars in host names (except for pct encoded)
+ */
 fn is_ireg_name_not_pct_encoded(c: char) -> bool {
-    is_iunreserved(c) || is_sub_delim(c)
+    is_iunreserved(c)
 }
 
 /// Parse host
@@ -114,10 +117,12 @@ fn parse_host(input: &str) -> IResult<&str, (&str, bool), CustomError<&str>> {
 }
 
 fn take_while_ireg(input: &str) -> IResult<&str, &str, CustomError<&str>> {
-    recognize(many0(alt((
+    let (input, result) = recognize(many0(alt((
         recognize(many1(take_while_pct_encoded)),
         take_while1(is_ireg_name_not_pct_encoded),
-    ))))(input)
+    ))))(input)?;
+
+    Ok((input, result.trim_end_matches('.')))
 }
 
 /// Parse the iauthority block