+++ tests/text_to_ast/desktop_set.rs
@@ -1,5 +1,5 @@
 use super::*;
-use deltachat_message_parser::parser::parse_desktop_set;
+use deltachat_message_parser::parser::{link_url::PunycodeWarning, parse_desktop_set};
 
 #[test]
 fn do_not_parse_markdown_elements() {
@@ -405,6 +405,34 @@ fn inline_link_do_not_eat_last_char_if_it_is_special() {
             destination: https_link_no_puny("https://delta.chat/page.hi", "delta.chat")
         }]
     );
+    assert_eq!(
+        parse_desktop_set("So here's the link: https://delta.cat/page."),
+        vec![
+            Text("So here's the link: "),
+            Link {
+                destination: https_link_no_puny("https://delta.cat/page", "delta.cat")
+            },
+            Text(".")
+        ]
+    );
+    assert_eq!(
+        parse_desktop_set("Here's a list of organizations which funded DC: https://delta.chat/en/help#how-are-delta-chat-developments-funded."),
+        vec![
+            Text("Here's a list of organizations which funded DC: "),
+            Link { destination: https_link_no_puny("https://delta.chat/en/help#how-are-delta-chat-developments-funded", "delta.chat")},
+            Text(".")
+        ]
+    );
+    assert_eq!(
+        parse_desktop_set("So here's the link: https://delta.cat/page, have fun"),
+        vec![
+            Text("So here's the link: "),
+            Link {
+                destination: https_link_no_puny("https://delta.cat/page", "delta.cat")
+            },
+            Text(", have fun")
+        ]
+    );
 }
 
 #[test]
@@ -417,6 +445,37 @@ fn labeled_link() {
                 "https://delta.chat/en/help?hi=5&e=4#section2.0",
                 "delta.chat"
             ),
+        }]
+    );
+}
+
+#[test]
+fn labeled_link_with_special_char_in_domain() {
+    assert_eq!(
+        parse_desktop_set("[munich](https://münchen.de)"),
+        vec![LabeledLink {
+            label: vec![Text("munich")],
+            destination: LinkDestination {
+                target: "https://münchen.de",
+                hostname: Some("münchen.de"),
+                punycode: Some(PunycodeWarning {
+                    original_hostname: "münchen.de".to_string(),
+                    ascii_hostname: "xn--mnchen-3ya.de".to_string(),
+                    punycode_encoded_url: "https://xn--mnchen-3ya.de".to_string()
+                }),
+                scheme: "https"
+            },
+        }]
+    );
+}
+
+#[test]
+fn labeled_link_domain_only() {
+    assert_eq!(
+        parse_desktop_set("[a link](https://delta.chat)"),
+        vec![LabeledLink {
+            label: vec![Text("a link")],
+            destination: https_link_no_puny("https://delta.chat", "delta.chat"),
         }]
     );
 }