+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - fix: parse fediverse addresses as text, so they are not mistaken for email addresses ([issue #82](https://github.com/deltachat/message-parser/issues/82))
 - fix: don't consume/eat the the exlaimation mark at the end of a link #85
+- fix: don't consume/eat various characters at the end of host names #96
 
 ## 0.11.0 - Bug fixes for Link Parsing